+++ app/routers/user_routes.py
@@ -1,16 +1,17 @@
-from fastapi import APIRouter, Depends, Query
+from fastapi import APIRouter, Depends, Query, HTTPException
 from schemas import UserCreate, UserRead, UserUpdate
 from sqlalchemy.ext.asyncio import AsyncSession
 from database import get_db
 from crud.crud_user import create_user, read_user, read_users, update_user, delete_user
+from auth import get_current_user
+from models import User
 from typing import List
-
 
 user_router = APIRouter(prefix="/users", tags=["Users"])
 
 
 @user_router.post("/", response_model=UserRead, status_code=201)
-async def add_user(user: UserCreate, db: AsyncSession = Depends(get_db)) -> UserRead:
+async def add_user(user: UserCreate, db: AsyncSession = Depends(get_db)):
     return await create_user(user, db)
 
 
@@ -28,10 +29,23 @@ async def get_users(
 
 
 @user_router.patch("/{user_id}", response_model=UserRead)
-async def patch_user(user_id: int, new_user: UserUpdate, db: AsyncSession = Depends(get_db)):
+async def patch_user(
+        user_id: int,
+        new_user: UserUpdate,
+        db: AsyncSession = Depends(get_db),
+        current_user: User = Depends(get_current_user)):
+    if user_id != current_user.user_id:
+        raise HTTPException(
+            status_code=404, detail="You can only update your own profile")
     return await update_user(user_id, new_user, db)
 
 
 @user_router.delete("/{user_id}", response_model=UserRead, status_code=200)
-async def remove_user(user_id: int, db: AsyncSession = Depends(get_db)):
+async def remove_user(
+        user_id: int,
+        db: AsyncSession = Depends(get_db),
+        current_user: User = Depends(get_current_user)):
+    if user_id != current_user.user_id:
+        raise HTTPException(
+            status_code=404, detail="You can only delete your own profile")
     return await delete_user(user_id, db)
